pyproject.toml
[project]
name = "codex-orch"
version = "0.1.0"
description = "CLI orchestrator for Codex multi-worker workflows"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
codex-orch = "codex_orch.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
